+++ charactersheet/charactersheet/viewmodels/dm/encounter_sections/maps_and_images_section/index.js
@@ -1,6 +1,19 @@
 'use strict';
+import ko from 'knockout'
+import Strophe from 'strophe'
 
-function MapsAndImagesSectionViewModel(parentEncounter) {
+import { Message,
+    MapOrImage } from 'charactersheet/models'
+import { PlayerPushModalViewModel, MapsAndImagesSection } from 'charactersheet/viewmodels/dm'
+import { ImageServiceManager,
+    PersistenceService,
+    ChatServiceManager,
+    XMPPService } from 'charactersheet/services/common'
+import { Notifications, CharacterManager, Utility } from 'charactersheet/utilities'
+
+import template from './index.html'
+
+export function MapsAndImagesSectionViewModel(parentEncounter) {
     var self = this;
 
     self.template = 'maps_and_images_section.tmpl';
@@ -280,3 +293,8 @@ function MapsAndImagesSectionViewModel(parentEncounter) {
         self._isConnectedToParty(chat.currentPartyNode != null);
     };
 }
+
+ko.components.register('maps-and-images-section', {
+  viewModel: MapsAndImagesSectionViewModel,
+  template: template
+})
