+++ charactersheet/charactersheet/viewmodels/dm/party_tab/index.js
@@ -1,7 +1,10 @@
 'use strict';
+import ko from 'knockout'
 
+import { PartyViewModel } from 'charactersheet/viewmodels/dm'
+import { ViewModelUtilities } from 'charactersheet/utilities'
 
-function PartyTabViewModel() {
+export function PartyTabViewModel() {
     var self = this;
 
     self.partyViewModel = ko.observable(new PartyViewModel());
